src/helpers.ts: reimplement analyseRoll.
This fixes a mistake in the original implementation. The rules specify
advantage and disadvantage in terms of better and worse results, which are not
necessarily identical to higher or lower values. For critical fails are always
worse than regular fails, even if the critical failure is a lower number than
the regular fail. This entails a fairly convoluted logic. I chose to implement
this by mapping the pairs of values on to integers, so that I can then compare
the results using simple greater-than and less-than operations. This isn't the
most graceful code, but it does pass all tests, an improvement over the
previous version.