{Dm,Player}Session.tsx: Specify ICE server.
The default PeerJS ICE server is in some way unreliable and causes catastrophic
failures in practice. Set the ICE server to a reliable public stun server to
negotiate connections. The server made explicit was already the default: what
this change does is to remove the implicit other default servers that are added
after this one.