Rewrite applyDamage to account for armor.
The logic now accounts for armor, but this required a major refactor of the
types of entities in the game. Where before there were Contractors, NPCs,
Monsters, and Characters, there are now only PlayerCharacters and
NonPlayerCharacters, both of which inherit from the same Character
interface. I believe that the gain in flexibility compensates for the loss in
precision. All the non-player characters in the game have a quasi-arbitrary
mixture of characteristics that make it difficult to define clear types within
them. Graphically we can continue to distinguish between different entities,
but ultimately that is a question of which data structure the entity is stored
in more than a question of the type of entity that it is.