Many files: implement damage with armor.
This was a huge refactor of the entire code base. The types are now
unrecognizable. I'm sure a few issues remain to be ironed out in the UX, but it
mostly works. Most of the problems that remain can be considered issues of
workflow and design rather than of logic. I'm sure, also, that the JSON schema
I ended up using for the weapons is an intense kludge. I'm still not happy with
how the wound types are juggled about, and any time you pass a list in to a
react element it warns you that it should have an id attribute.