Add wound argument to calls to rollDamage.
The rollDamage function shows a message with the type of damage dealt and a
button to take damage. Therefore the component must know the type of damage
dealt. But since wound type is an attribute of the weapon or attack and not of
the damage object itself (the latter gives instructions on how to roll the
amount of damage), this function needs to be passed the wound type explicitly.