Disable automatic application of wounds on taking damage.
For now this logic is too buggy and clunky to be useful. Players will be
responsible for rolling their own wounds when they get them. There are some
problems in automating this that spring all the way from the data structure
used to represent wounds. The Laser Cutter, which rolls *either* Bleeding or
Gore, and the Vibechete, which rolls *both* Bleeding and Gore are not
meaningfully distinguished in their representation of wounds, so it's not
possible for the logic to seperate the cases except by hard-coding special
exceptions for those particular weapons. This will require a rethink to
properly generalize the way to do this. Similarly, it's an open question
how to handle rolling wounds for NPCs, since a lot of the wound effects aren't
defined for them.