src/Services/damageServices.ts: change signature of applyWounds.
Add a number argument, because for a minute there I really thought I was going
to automate applying wounds. For now this argument isn't used, and the wounds
are always rolled manually. Similarly, remove the optional null from log, since
it causes more trouble than it's worth, and one can always pass a nonce
function (`() => {}`) to satisfy the type checker.