Revert 565aba5: correct roll mode in applyWounds.
On closer examination, this is an irregularity in the semantics of the
rolls. "Better" and "worse" in the context of wound rolls are from the
perspective of the person *dealing* the damage, who wants to do as much as
possible and therefore prefers to roll high, and not the person *receiving* the
damage, who would prefer to roll low. This makes sense of why the more powerful
weapons roll wounds at an advantage, and the weakers ones at a
disadvantage. This could probably be spelled out more clealy in the rule book,
but I'm satisfied that this is the intended semantics.